requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.1.0
django-phonenumber-field==7.1.0
django-ssl-smtp==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumberslite==8.13.15
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
